WidgetMarket.Users (widget_market v0.2.0)

Functions

authenticate(params)

Callback implementation for Pow.Context.authenticate/1.

create(params)

Callback implementation for Pow.Context.create/1.

credit(seller_id, widget_price)

subtracts the market fee from the widget price and credits the difference to the seller

debit(buyer_id, widget_price)

subtracts the the widget price from the buyer if the buyer has no credit, an error tuple is returned

delete(user)

Callback implementation for Pow.Context.delete/1.

get_by(clauses)

Callback implementation for Pow.Context.get_by/1.

get_user(id)

Gets a single user.

Raises Ecto.NoResultsError if the User does not exist.

Examples

iex> get_user(123)
%User{}

iex> get_user(456)
** (Ecto.NoResultsError)

list_users()

Gets all users

Examples

iex> list_users()
[%User{}]

pow_authenticate(params)

pow_create(params)

pow_delete(user)

pow_get_by(clauses)

pow_update(user, params)

update(user, params)

Callback implementation for Pow.Context.update/2.