WidgetMarketWeb.WidgetController (widget_market v0.2.0)

This module contains the controller actions for widget manipulation

Functions

create(conn, map)

Creates a new widget

delete(conn, map)

Deletes a widget. Only the widget owner is allowed to do this.

edit(conn, map)

Yields a form for updating a widget

index(conn, params)

Returns the list of widgets.

new(conn, params)

Yields a form for creating a new widget

purchase(conn, map)

Purchases a widget. Creates a marketplace transaction

show(conn, map)

Displays a widget

update(conn, map)

Updates a widget. Only the widget owner is allowed to do this.